+++ packages/python/qsu/math/__init__.py
@@ -3,6 +3,7 @@ from .div import div
 from .mul import mul
 from .numPick import numPick
 from .numUnique import numUnique
+from .round import round
 from .sub import sub
 from .sum import sum
 
@@ -12,6 +13,7 @@ __all__ = [
 	'mul',
 	'numPick',
 	'numUnique',
+	'round',
 	'sub',
 	'sum',
 ]